Just committing to make sure I don't lose my progress. Not able to compile at this time!!!

+++ src/main/java/debug/Debugger.java
@@ -213,7 +213,7 @@ public class Debugger{
         });
     }
 
-    public void update(short opcode){
+    public void update(int opcode){
         //NOTE::Checking if we are stepping through instructions
         if(stepThroughInstructions){
             while(!stepThroughNextInstruction)
